moved hubpostcreate and addhubtomodal to sdk and refactored for comportability
