+++ js/sdk/rollup.config.js
@@ -59,6 +59,7 @@ export default [
       SwapModal: 'src/components/SwapModal.js',
       IdentityVerification: 'src/components/IdentityVerification.js',
       AddToHubModal: 'src/components/AddToHubModal.js',
+      HubPostCreate: 'src/components/HubPostCreate',
     },
     output: [
       {
