+++ js/sdk/src/components/HubPostCreateForm.js
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect } from 'react'
 import { styled } from '@mui/material/styles'
-import Release from '@nina-protocol/nina-internal-sdk/esm/Release'
 import { formatPlaceholder } from '@nina-protocol/nina-internal-sdk/esm/utils'
+import Release from '@nina-protocol/nina-internal-sdk/esm/Release'
 import { withFormik, Form, Field } from 'formik'
 import Typography from '@mui/material/Typography'
 import TextField from '@mui/material/TextField'
@@ -14,14 +14,12 @@ const QuillEditor = dynamic(
   () => import('@nina-protocol/nina-internal-sdk/esm/QuillEditor'),
   { ssr: false }
 )
+
 const HubPostCreateForm = ({
-  field,
-  form,
   values,
   onChange,
   errors,
   touched,
-  setFieldValue,
   update,
   hubData,
   postCreated,
@@ -45,16 +43,18 @@ const HubPostCreateForm = ({
 
   return (
     <Root>
+      {update && <Typography>Updating {hubData.name}</Typography>}
       <Form style={{ padding: '0 15px', height: '100%' }}>
         <Field name="title">
           {(props) => (
-            <Box>
+            <Box mb={1}>
               <TextField
                 className="formField"
                 variant="standard"
+                fullWidth
                 label={
                   formatPlaceholder(props.field.name) +
-                  (update ? ` (${hubData.data.title})` : '')
+                  (update ? ` (${hubData.json.title})` : '')
                 }
                 size="small"
                 InputLabelProps={touched.title ? { shrink: true } : ''}
@@ -70,11 +70,7 @@ const HubPostCreateForm = ({
         <Field name="body">
           {(props) => (
             <Box>
-              <QuillEditor
-                formikProps={props}
-                type={'post'}
-                postCreated={postCreated}
-              />
+              <QuillEditor formikProps={props} type={'post'} />
             </Box>
           )}
         </Field>
@@ -82,7 +78,7 @@ const HubPostCreateForm = ({
         {!preloadedRelease && (
           <Field name="reference">
             {(props) => (
-              <FormControl fullWidth style={{ marginTop: '50px' }}>
+              <FormControl fullWidth>
                 <Select
                   className="formField"
                   value={props.field.value}
@@ -112,7 +108,7 @@ const HubPostCreateForm = ({
         )}
 
         {preloadedRelease && (
-          <Typography mt={3}>
+          <Typography mt={6}>
             <i>{releaseState.metadata[preloadedRelease].name}</i> will be
             associated with this post
           </Typography>
@@ -122,7 +118,7 @@ const HubPostCreateForm = ({
   )
 }
 
-const Root = styled('div')(({ theme }) => ({
+const Root = styled('div')(() => ({
   width: '100%',
   height: '100%',
 }))
