+++ js/sdk/src/index.js
@@ -36,3 +36,4 @@ export { default as NoSolWarning } from './components/NoSolWarning'
 export { default as SwapModal } from './components/SwapModal'
 export { default as IndentityVerification } from './components/IndentityVerification'
 export { default as AddToHubModal } from './components/AddToHubModal'
+export { default as HubPostCreate } from './components/HubPostCreate'
